Upgrade the Schema Validator :heart:
Added the possibility to specify some constraints on number fields
and on string fields.

Strings and numbers can have a min/max length.

We can check that strings are correct emails, that they are in lowercase,
or in uppercase or that they are trimmed.

Evidently, all these cases can be combined.

By default, `lowercase`, `uppercase` and `trim` methods modify
the string to make it match. This option can be disabled by
passing `false` to the method.

What is super cool is the `alternatives` static method on Validator
class : it permits to check a value against multiple schemas !